% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mne-package.R
\docType{package}
\name{mne-package}
\alias{mne}
\alias{mne-package}
\title{mne: Fast Access to MNE-Python From Within R}
\description{
This package is documents how to use MNE-Python from
    within R using the retiuclate package. As needed, a small set of functions will be added
    which facilitate the interaction with MNE-Python.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mne-tools/mne-r}
  \item Report bugs at \url{https://github.com/mne-tools/mne-r}
}

}
\author{
\strong{Maintainer}: Denis Engemann \email{denis-alexander.engemannt@inria.fr} (0000-0002-7223-1014)

}
\keyword{internal}
