% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mne.r
\name{get_data_frame}
\alias{get_data_frame}
\title{Get data from MNE into R data.frame}
\usage{
get_data_frame(inst, picks = NULL, index = NULL, scaling_time = 1000,
  scalings = NULL, copy = T, start = NULL, stop = NULL,
  long_format = T)
}
\arguments{
\item{inst}{An instance of MNE data containsers, e.g,
\code{mne$Epochs}, \code{mne$io$Raw}, \code{mne$Evoked}.}

\item{picks}{a zero-indexed integer array, a string, list, slice or
None.}

\item{index}{The columns to be uesed as pandas index. tuple of str
or None}

\item{scaling_time}{Scaling to be applied to time units. Float.}

\item{scalings}{Scaling to be applied to the channels picked.}

\item{copy}{Whether to make a copy of the data.}

\item{start}{If it is a Raw object, this defines a starting index
for creating the dataframe from a slice. The times will be
interpolated from the index and the sampling rate of the
signal. Int or None.}

\item{stop}{If it is a Raw object, this defines a stop index for
creating the dataframe from a slice. The times will be
interpolated from the index and the sampling rate of the
signal. Int or None.}

\item{long_format}{If True, the dataframe is returned in long
format where each row is one observation of the signal at a
unique coordinate of channels, time points, epochs and
conditions. The number of factors depends on the data
container. For convenience, a ch_type column is added when
using this option that will facilitate subsetting the
resulting dataframe. Defaults to False.}
}
\value{
Returns a data.frame. The layout depends on the options.
        and the type of instance.
}
\description{
\code{get_data_frame} helps importing MNE data structures
}
\details{
The code will call the `to_data_frame_method` of the MNE instanc
passed. This makes it easy to use MNE-objects in R. The below
Type definitions are expressed in Python types. Please refer to
the reticulate documentation to learn about R-to-Python conversion
rules.
Note that this function requires, next ot MNE, a working Pandas
installation
}
